+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Not release
+## 5.1.1 - 2019-06-20
 
 ### Added
 - Check method for Alert
+
+### Deprecated
+- ItemMenu method `show` use `canSee`
+
+### Fixed
+- Relation field null search string [894](https://github.com/orchidsoftware/platform/pull/894)
 
 ## 5.1.0 - 2019-06-17